+++ src/osyris/io/hilbert.py
@@ -2,12 +2,10 @@
 # Copyright (c) 2021 Osyris contributors (https://github.com/nvaytet/osyris)
 
 import numpy as np
+from ..core import Array
 
 
 def _read_bound_key(infofile, ncpu):
-    """
-    Read hilbert bound key.
-    """
     with open(infofile) as f:
         content = f.readlines()
     starting_line = None
@@ -21,8 +19,6 @@ def _read_bound_key(infofile, ncpu):
             line = content[starting_line + n].split()
             bound_key.append(int(float(line[1])))
         bound_key.append(int(float(content[starting_line + ncpu - 1].split()[2])))
-
-    # return np.array(bound_key, dtype=float)
     return bound_key
 
 
@@ -32,52 +28,10 @@ def _btest(i, pos):
 
 def _hilbert3d(x, y, z, bit_length):
 
-    # logical,dimension(0:3*bit_length-1)::i_bit_mask
-    # logical,dimension(0:1*bit_length-1)::x_bit_mask,y_bit_mask,z_bit_mask
-    # integer,dimension(0:7,0:1,0:11)::state_diagram
-
-    # integer::i,ip,cstate,nstate,b0,b1,b2,sdigit,hdigit
-
-    # if(bit_length>bit_size(bit_length))then
-    #  write(*,*)'Maximum bit length=',bit_size(bit_length)
-    #  write(*,*)'stop in hilbert3d'
-    #  stop
-    # endif
-    # print('============')
-    # print(x, y, z, bit_length)
-
     i_bit_mask = np.zeros(3 * bit_length, dtype=bool)
     x_bit_mask = np.zeros(bit_length, dtype=bool)
     y_bit_mask = np.zeros_like(x_bit_mask)
     z_bit_mask = np.zeros_like(x_bit_mask)
-    # logical,dimension(0:3*bit_length-1)::i_bit_mask
-    # logical,dimension(0:1*bit_length-1)::x_bit_mask,y_bit_mask,z_bit_mask
-
-    # state_diagram = RESHAPE( (/   1, 2, 3, 2, 4, 5, 3, 5,&
-    #    &   0, 1, 3, 2, 7, 6, 4, 5,&
-    #    &   2, 6, 0, 7, 8, 8, 0, 7,&
-    #    &   0, 7, 1, 6, 3, 4, 2, 5,&
-    #    &   0, 9,10, 9, 1, 1,11,11,&
-    #    &   0, 3, 7, 4, 1, 2, 6, 5,&
-    #    &   6, 0, 6,11, 9, 0, 9, 8,&
-    #    &   2, 3, 1, 0, 5, 4, 6, 7,&
-    #    &  11,11, 0, 7, 5, 9, 0, 7,&
-    #    &   4, 3, 5, 2, 7, 0, 6, 1,&
-    #    &   4, 4, 8, 8, 0, 6,10, 6,&
-    #    &   6, 5, 1, 2, 7, 4, 0, 3,&
-    #    &   5, 7, 5, 3, 1, 1,11,11,&
-    #    &   4, 7, 3, 0, 5, 6, 2, 1,&
-    #    &   6, 1, 6,10, 9, 4, 9,10,&
-    #    &   6, 7, 5, 4, 1, 0, 2, 3,&
-    #    &  10, 3, 1, 1,10, 3, 5, 9,&
-    #    &   2, 5, 3, 4, 1, 6, 0, 7,&
-    #    &   4, 4, 8, 8, 2, 7, 2, 3,&
-    #    &   2, 1, 5, 6, 3, 0, 4, 7,&
-    #    &   7, 2,11, 2, 7, 5, 8, 5,&
-    #    &   4, 5, 7, 6, 3, 2, 0, 1,&
-    #    &  10, 3, 2, 6,10, 3, 4, 4,&
-    #    &   6, 1, 7, 0, 5, 2, 4, 3 /), &
-    #    & (/8 ,2, 12 /) )
 
     state_diagram = np.array([
         1, 2, 3, 2, 4, 5, 3, 5, 0, 1, 3, 2, 7, 6, 4, 5, 2, 6, 0, 7, 8, 8, 0, 7, 0, 7, 1,
@@ -90,26 +44,19 @@ def _hilbert3d(x, y, z, bit_length):
         4, 4, 6, 1, 7, 0, 5, 2, 4, 3
     ]).reshape((8, 2, 12), order='F')
 
-    # convert to binary
+    # Convert to binary
     for i in range(bit_length):
         x_bit_mask[i] = _btest(x, i)
         y_bit_mask[i] = _btest(y, i)
         z_bit_mask[i] = _btest(z, i)
 
-    # print(x_bit_mask)
-    # print(y_bit_mask)
-    # print(z_bit_mask)
-
-    # interleave bits
+    # Interleave bits
     for i in range(bit_length):
         i_bit_mask[3 * i + 2] = x_bit_mask[i]
         i_bit_mask[3 * i + 1] = y_bit_mask[i]
         i_bit_mask[3 * i] = z_bit_mask[i]
 
-    # print(i_bit_mask)
-
-    # build Hilbert ordering using state diagram
-    # print('cstate')
+    # Build Hilbert ordering using state diagram
     cstate = 0
     for i in range(bit_length - 1, -1, -1):
         b2 = 0
@@ -128,28 +75,18 @@ def _hilbert3d(x, y, z, bit_length):
         i_bit_mask[3 * i + 1] = _btest(hdigit, 1)
         i_bit_mask[3 * i] = _btest(hdigit, 0)
         cstate = nstate
-    #     print(i)
-    #     print(b2, b1, b0)
-    #     print(sdigit, nstate, hdigit)
-    #     print(i_bit_mask)
-    #     print(cstate)
-    # print('end cstate ---------')
 
-    # save Hilbert key as double precision real ??
     order = 0
     for i in range(3 * bit_length):
         b0 = 0
         if i_bit_mask[i]:
             b0 = 1
         order = order + b0 * (2**i)
-
-    # print(order, type(order))
     return order
 
 
-def hilbert_cpu_list(bounding_box, lmax, levelmax, infofile, ncpu, ndim):
+def _get_cpu_list(bounding_box, lmax, levelmax, infofile, ncpu, ndim):
 
-    # Read bound key
     bound_key = _read_bound_key(infofile=infofile, ncpu=ncpu)
 
     xmin = bounding_box["xmin"]
@@ -158,9 +95,7 @@ def hilbert_cpu_list(bounding_box, lmax, levelmax, infofile, ncpu, ndim):
     ymax = bounding_box["ymax"]
     zmin = bounding_box["zmin"]
     zmax = bounding_box["zmax"]
-
     dmax = max(xmax - xmin, ymax - ymin, zmax - zmin)
-
     for ilevel in range(1, lmax + 1):
         dx = 0.5**ilevel
         if dx < dmax:
@@ -182,16 +117,6 @@ def hilbert_cpu_list(bounding_box, lmax, levelmax, infofile, ncpu, ndim):
         jmax = jmin + 1
         kmin = int(zmin * maxdom)
         kmax = kmin + 1
-    #     imin = int(xmin * dble(maxdom))
-    #     imax = imin + 1
-    #     jmin = int(ymin * dble(maxdom))
-    #     jmax = jmin + 1
-    #     kmin = int(zmin * dble(maxdom))
-    #     kmax = kmin + 1
-    # endif
-
-    # print(dmax, bit_length, maxdom, lmin)
-    # print(imin, imax, jmin, jmax, kmin, kmax)
 
     dkey = (2**(levelmax + 1) // maxdom)**ndim
     ndom = 1
@@ -201,45 +126,17 @@ def hilbert_cpu_list(bounding_box, lmax, levelmax, infofile, ncpu, ndim):
     jdom = [jmin, jmin, jmax, jmax] * 2
     kdom = [kmin] * 4 + [kmax] * 4
 
-    # print(dkey, ndim)
-    # print(idom)
-    # print(jdom)
-    # print(kdom)
-    # idom(1)=imin; idom(2)=imax
-    # idom(3)=imin; idom(4)=imax
-    # idom(5)=imin; idom(6)=imax
-    # idom(7)=imin; idom(8)=imax
-    # jdom(1)=jmin; jdom(2)=jmin
-    # jdom(3)=jmax; jdom(4)=jmax
-    # jdom(5)=jmin; jdom(6)=jmin
-    # jdom(7)=jmax; jdom(8)=jmax
-    # kdom(1)=kmin; kdom(2)=kmin
-    # kdom(3)=kmin; kdom(4)=kmin
-    # kdom(5)=kmax; kdom(6)=kmax
-    # kdom(7)=kmax; kdom(8)=kmax
-
-    # do i=1,ndom
-
     bounding_min = [0, 0, 0, 0, 0, 0, 0, 0]
     bounding_max = [0, 0, 0, 0, 0, 0, 0, 0]
     bounding = None
     for i in range(ndom):
-        # print(i)
         if bit_length > 0:
             bounding = _hilbert3d(idom[i], jdom[i], kdom[i], bit_length)
             order_min = bounding
-            # print('hilbert', order_min)
         else:
             order_min = 0
-            # print('zero', order_min)
-        # print(order_min, bounding)
         bounding_min[i] = order_min * dkey
         bounding_max[i] = (order_min + 1) * dkey
-    # return
-
-    # print(bounding_min)
-    # print(bounding_max)
-    # return
 
     cpu_min = [0, 0, 0, 0, 0, 0, 0, 0]
     cpu_max = [0, 0, 0, 0, 0, 0, 0, 0]
@@ -253,20 +150,48 @@ def hilbert_cpu_list(bounding_box, lmax, levelmax, infofile, ncpu, ndim):
                     and (bound_key[impi + 1] >= bounding_max[i])):
                 cpu_max[i] = impi
 
-    # ncpu_read=0
     cpu_list = []
     for i in range(ndom):
         for j in range(cpu_min[i], cpu_max[i] + 1):
             if j + 1 not in cpu_list:
                 cpu_list.append(j + 1)
 
-    # print(cpu_list)
     return cpu_list
 
 
-# else
-#    ncpu_read=ncpu
-#    do j=1,ncpu
-#       cpu_list(j)=j
-#    end do
-# end  if
+def hilbert_cpu_list(meta, scaling, select, infofile):
+    if meta["ordering type"] == "hilbert":
+        bounding_box = {
+            "xmin": 0,
+            "xmax": 1,
+            "ymin": 0,
+            "ymax": 1,
+            "zmin": 0,
+            "zmax": 1
+        }
+        # Make an array of cell centers according to lmax
+        box_size = (meta["boxlen"] * scaling).magnitude
+        ncells = 2**min(meta["levelmax"], 18)  # limit to 262000 cells
+        half_dxmin = 0.5 * box_size / ncells
+        xyz_centers = Array(values=np.linspace(half_dxmin, box_size - half_dxmin,
+                                               ncells),
+                            unit=1.0 * scaling.units)
+        new_bbox = False
+        for c in "xyz":
+            if c in select:
+                new_bbox = True
+                func_test = select[c](xyz_centers)
+                inds = np.argwhere(func_test).ravel()
+                start = xyz_centers[inds.min()] - (half_dxmin * scaling.units)
+                end = xyz_centers[inds.max()] + (half_dxmin * scaling.units)
+                bounding_box["{}min".format(c)] = start._array / box_size
+                bounding_box["{}max".format(c)] = end._array / box_size
+                select["xyz_{}".format(c)] = select.pop(c)
+
+        if new_bbox:
+            return _get_cpu_list(bounding_box=bounding_box,
+                                 lmax=meta["lmax"],
+                                 levelmax=meta["levelmax"],
+                                 infofile=infofile,
+                                 ncpu=meta["ncpu"],
+                                 ndim=meta["ndim"])
